- fix w4 warnings - add convenience macros for config_db(CFG_GET_SYS_VAL) - VFSUtil::EnumDirEnts now uses flags instead of bool recursive - UNUSED() for params, UNUSED2 (<- need better name) for variables - config.h defines must be tested with #if (always defined) -> allows detecting misspellings thanks to compiler warnings - replace debug_assert(0) with debug_warn (its sole purpose) - replace ScriptingHost::ValueToInt et al with ToPrimitive - use nommgr.h to disable both mmgr and VC debug heap
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2585 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/ModelDef.cpp
@@ -107,11 +107,11 @@ void CModelDef::Save(const char* filename,const CModelDef* mdef)
 	CFilePacker packer(FILE_VERSION, "PSMD");
 
 	// pack everything up
-	u32 numVertices=mdef->GetNumVertices();
+	u32 numVertices=(u32)mdef->GetNumVertices();
 	packer.PackRaw(&numVertices,sizeof(numVertices));
 	packer.PackRaw(mdef->GetVertices(),sizeof(SModelVertex)*numVertices);
 	
-	u32 numFaces=mdef->GetNumFaces();
+	u32 numFaces=(u32)mdef->GetNumFaces();
 	packer.PackRaw(&numFaces,sizeof(numFaces));
 	packer.PackRaw(mdef->GetFaces(),sizeof(SModelFace)*numFaces);
